refactor(data): move column helpers to data/columns/ (refs #298)
