+++ src/components/DataTablePanel.tsx
@@ -25,12 +25,12 @@ import { useApplyTransformation } from 'hooks/useApplyTransformation';
 import React, { useEffect, useMemo, useRef, useState } from 'react';
 import { DatatableOptions } from 'types';
 import { BuildColumnDefs, ConvertDataFrameToDataTableFormat } from 'data/dataHelpers';
-import { ApplyColumnWidthHints } from 'data/columnWidthHints';
+import { ApplyColumnWidthHints } from 'data/columns/columnWidthHints';
 import { buildSearchBarLayout } from 'data/buildSearchBarLayout';
 import { datatableThemedStyles } from './styles';
 import { GetDataTransformerID } from 'data/transformations';
 import { DTColumnType } from 'data/types';
-import { ApplyColumnAliases } from 'data/columnAliasing';
+import { ApplyColumnAliases } from 'data/columns/columnAliasing';
 
 interface Props extends PanelProps<DatatableOptions> { }
